+++ ui/dashboards/src/context/index.ts
@@ -14,5 +14,4 @@
 export * from './DashboardProvider';
 export * from './DatasourceStoreProvider';
 export * from './TemplateVariableProvider';
-export * from './TimeRangeProvider';
 export * from './useDashboard';